requirements.txt
RPi.bme280==0.2.4
smbus2==0.5.0
kafka-python==2.0.2